fix(function calls): add proper support for sending function messages in stream
Prior to this PR, it was not possible to send the assistant message that selected
a function to OpenAI via the gpt-4 plugin. The full protocol appears to require
that the assistant message be sent.

I suspect this might matter for chained function invocations, etc. This PR sets
the stage for a unified `History` object (for streaming) that contains the full
set of agent<-->llm interactions.

A slight refactor of the functions labeling is supported. This is from two tags
(role and name) to a single tag (role with a value of the name). The legacy pattern
is still supported for backwards compatibility.